ivy: statically distinguish local from global variables
Following Python's lead, a local variable is one that is written
before being read in the program execution order, and
a global variable is one that is read before being written.

The statement 'x;' serves to declare x as a global, assuming
x has not been assigned yet. We could always define an operator too.

Because we have this analysis, we can also count the number
of local variables, assign them indices, and use a slice instead of a map
to hold them, so this change does that as well.

FOR DISCUSSION. It is unclear whether this is the right rule,
and probably there is still polishing to do. It is unclear where
doc changes should be applied, for example.